Substituted atoms for words
Replaced everything that the lexer recognizes with words.
