+++ Newlanguage/newGrammar
@@ -1,0 +1,25 @@
+// '|' means OR, and '*' means 0 or more of what was before.
+expr    : KEYWORD:"LET ITEM" IDENTIFIER EQ expr
+        : ((plus|minus) term and term)* //An expression is 0 or more terms added or subtracted with each other.
+
+term    : ((mul|div) factor and factor)* //A term is 0 or more factors multiplied or divided by each other.
+
+factor  : (PLUS|MINUS) factor
+        : power
+
+power   : atom ("to the power of" factor)*
+
+
+atom    : INT|FLOAT|IDENTIFIER //a factor is a number, either an int or a float. or an identifier which references another value.
+        : LPAREN expr RPAREN
+
+Math operatoions
+
+"the [sum/difference/product/quotient] of [term] and [term]."
+
+Combining 2 strings
+"'string one here', 'string two here'"
+double quotations instead of single
+
+declaring and referencing varaibles.
+"Let item [variable name] be [value]"
